requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
google-generativeai==0.7.2
python-dotenv
waitress==2.1.2